refactor(core): move logging_protocol to core; update imports (P16 step 3)
