+++ ml_playground/checkpoint.py
@@ -9,7 +9,7 @@ import torch
 from torch.serialization import pickle as _torch_pickle  # type: ignore[attr-defined]
 
 from ml_playground.core.error_handling import CheckpointError, CheckpointLoadError
-from ml_playground.logging_protocol import LoggerLike
+from ml_playground.core.logging_protocol import LoggerLike
 
 TorchUnpicklingError = _torch_pickle.UnpicklingError  # type: ignore[attr-defined]
