+++ ml_playground/data_pipeline/transforms/io.py
@@ -10,7 +10,7 @@ import numpy as np
 from ml_playground._file_state import diff_file_states, snapshot_file_states
 from ml_playground.configuration.models import DataConfig
 from ml_playground.core.error_handling import DataError
-from ml_playground.logging_protocol import LoggerLike
+from ml_playground.core.logging_protocol import LoggerLike
 from ml_playground.data_pipeline.transforms.tokenization import coerce_tokenizer_type
 from ml_playground.core.tokenizer import create_tokenizer
 from ml_playground.tokenizer_protocol import Tokenizer
